% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_seasonalbreaks.R
\name{seasonalbreaks}
\alias{seasonalbreaks}
\title{Title}
\usage{
seasonalbreaks(
  y,
  level = 1,
  slope = 1,
  noise = 1,
  seasonal = c("Trigonometric", "Dummy", "Crude", "HarrisonStevens", "Fixed", "Unused"),
  X = NULL,
  X.td = NULL
)
}
\arguments{
\item{y}{}

\item{level}{-1 = no level, 0 = fixed level, 1 = sotchastic level}

\item{slope}{}

\item{noise}{}

\item{seasonal}{Seasonal model}

\item{X}{Regression variables (same length as y) or NULL}

\item{X.td}{Specification for trading days clustering. 
Contains thr group id fo Mondays... Sundays 
(for instance (1,1,1,1,1,0,0) for week days or (1,1,1,1,1,2,0) for week.saturdays/Sundays variables).
Contrasts are used. Can be NULL}
}
\value{

}
\description{
Title
}
