% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_stsoutliers.R
\name{sts.outliers}
\alias{sts.outliers}
\title{Title}
\usage{
sts.outliers(
  y,
  level = 1,
  slope = 1,
  cycle = -1,
  noise = 1,
  seasonal = c("Trigonometric", "Dummy", "Crude", "HarrisonStevens", "Fixed", "Unused"),
  X = NULL,
  X.td = NULL,
  cv = 0,
  tcv = 0,
  estimation.forward = c("Score", "Point", "Full"),
  estimation.backward = c("Point", "Score", "Full")
)
}
\arguments{
\item{y}{}

\item{level}{}

\item{slope}{}

\item{cycle}{}

\item{noise}{}

\item{seasonal}{}

\item{X}{}

\item{X.td}{}

\item{cv}{}

\item{tcv}{}

\item{estimation.forward}{}

\item{estimation.backward}{}
}
\value{

}
\description{
Title
}
