% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_sts.R
\name{sts}
\alias{sts}
\title{Structural time series}
\usage{
sts(
  y,
  level = 1,
  slope = 1,
  cycle = -1,
  noise = 1,
  seasonal = c("Trigonometric", "Dummy", "Crude", "HarrisonStevens", "Fixed", "Unused")
)
}
\arguments{
\item{y}{}

\item{level}{}

\item{slope}{}

\item{cycle}{}

\item{noise}{}

\item{seasonal}{}
}
\value{

}
\description{
Structural time series
}
