% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_stsoutliers.R
\name{tradingdays}
\alias{tradingdays}
\title{Title}
\usage{
tradingdays(clustering, freq, startyear, startperiod = 1, len, contrasts = T)
}
\arguments{
\item{clustering}{}

\item{freq}{}

\item{startyear}{}

\item{startperiod}{}

\item{len}{}
}
\value{

}
\description{
Title
}
