% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jd3_tramo.R
\name{tramooutliers}
\alias{tramooutliers}
\title{Title}
\usage{
tramooutliers(
  y,
  order = c(0L, 1L, 1L),
  seasonal = c(0L, 1L, 1L),
  mean = F,
  X = NULL,
  X.td = NULL,
  ao = T,
  ls = T,
  so = F,
  tc = F,
  cv = 0,
  ml = F
)
}
\arguments{
\item{y}{}

\item{order}{}

\item{seasonal}{}

\item{mean}{}

\item{X}{}

\item{X.td}{}

\item{ao}{}

\item{ls}{}

\item{so}{}

\item{tc}{}

\item{cv}{}

\item{ml}{}
}
\value{

}
\description{
Title
}
